ai-engine/requirements.txt
langgraph>=0.2.0
qdrant-client>=1.10.0
trafilatura
aiohttp>=3.9.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
# pydantic>=2.5.0>=1.10.0
pydantic>=2.5.0
feedparser>=6.0.10
jinja2>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
